backend/requirements.txt
Django==5.0.4
asgiref==3.8.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
